+++ client/src/main/java/net/oneandone/stool/client/Project.java
@@ -15,11 +15,9 @@
  */
 package net.oneandone.stool.client;
 
-import net.oneandone.inline.ArgumentException;
 import net.oneandone.sushi.fs.file.FileNode;
 import net.oneandone.sushi.launcher.Failure;
 import net.oneandone.sushi.launcher.Launcher;
-import net.oneandone.sushi.util.Strings;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -185,118 +183,5 @@ public class Project {
             addWars(project, lazyWars);
         }
         return lazyWars;
-    }
-
-    //-- stage name
-
-    /**
-     * The stage name has to be a valid domain name because is used as part of the application url.
-     * See http://tools.ietf.org/html/rfc1035 section 2.3.1.
-     */
-    public static void checkName(String name) {
-        char c;
-
-        if (name.isEmpty()) {
-            throw new ArgumentException("empty stage name is not allowed");
-        }
-        if (name.length() > 30) {
-            //ITCA does not accept too long commonNames
-            throw new ArgumentException("Stage Name is too long. Please take a shorter one.");
-        }
-        if (!isLetter(name.charAt(0))) {
-            throw new ArgumentException("stage name does not start with a letter");
-        }
-        for (int i = 1; i < name.length(); i++) {
-            c = name.charAt(i);
-            if (!isValidStageNameChar(c)) {
-                throw new ArgumentException("stage name contains illegal character: " + c);
-            }
-        }
-    }
-
-    public static boolean isValidStageNameChar(char c) {
-        return isLetter(c) || isDigit(c) || c == '-' || c == '.';
-    }
-    // cannot use Character.is... because we check ascii only
-    private static boolean isLetter(char c) {
-        return (c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z');
-    }
-    // cannot use Character.is... because we check ascii only
-    private static boolean isDigit(char c) {
-        return c >= '0' && c <= '9';
-    }
-
-    public static String nameForUrl(String url) {
-        if (url.startsWith("gav:")) {
-            return nameForGavUrl(url);
-        } else if (url.startsWith("file:")) {
-            return nameForFileUrl(url);
-        } else {
-            return nameForSvnOrGitUrl(url);
-        }
-    }
-
-    private static String nameForGavUrl(String url) {
-        int end;
-        int start;
-
-        url = one(url);
-        end = url.lastIndexOf(':');
-        if (end == -1) {
-            return "stage";
-        }
-        start = url.lastIndexOf(':', end - 1);
-        if (start == -1) {
-            return "stage";
-        }
-        return url.substring(start + 1, end);
-    }
-
-    private static String nameForFileUrl(String url) {
-        int idx;
-
-        url = one(url);
-        idx = url.lastIndexOf('/');
-        if (idx == -1) {
-            return "idx";
-        }
-        url = url.substring(idx + 1);
-        idx = url.lastIndexOf('.');
-        if (idx == -1) {
-            return url;
-        } else {
-            return url.substring(0, idx);
-        }
-    }
-
-    private static String one(String url) {
-        int end;
-
-        end = url.lastIndexOf(',');
-        if (end != -1) {
-            url = url.substring(0, end);
-        }
-        end = url.lastIndexOf('=');
-        if (end != -1) {
-            url = url.substring(0, end);
-        }
-        return url;
-    }
-
-    private static String nameForSvnOrGitUrl(String url) {
-        String result;
-        int idx;
-
-        result = Strings.removeRightOpt(url, "/");
-        idx = result.indexOf(':');
-        if (idx != -1) {
-            // strip protocol - important vor gav stages
-            result = result.substring(idx + 1);
-        }
-        result = Strings.removeRightOpt(result, "/trunk");
-        idx = result.lastIndexOf('/');
-        result = result.substring(idx + 1); // ok for -1
-        result = Strings.removeRightOpt(result, ".git");
-        return result.isEmpty() ? "stage" : result;
     }
 }